chore(tests): move remaining valid tests next to actual files
